requirements.txt
yfinance>=0.2.31
hmmlearn>=0.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
plotly>=5.17.0
ta>=0.10.2
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
